[MOV] product_expiry: split file product_expiry.py
