+++ addons/product_expiry/__openerp__.py
@@ -19,6 +19,8 @@ Following dates can be tracked:
 Also implements the removal strategy First Expiry First Out (FEFO) widely used, for example, in food industries.
 """,
     'data': ['views/product_expiry_view.xml',
+             'views/product_template_views.xml',
+             'views/stock_quant_views.xml',
              'data/product_expiry_data.xml'],
     'auto_install': False,
     'installable': True,
