+++ addons/product_expiry/models/__init__.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-import product_expiry
+import  production_lot
+import stock_quant
+import product_product
