+++ addons/product_expiry/models/production_lot.py
@@ -1,0 +1,52 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+import datetime
+
+import openerp
+from openerp.osv import fields, osv
+
+
+class stock_production_lot(osv.osv):
+    _inherit = 'stock.production.lot'
+
+    def _get_date(dtype):
+        """Return a function to compute the limit date for this type"""
+        def calc_date(self, cr, uid, context=None):
+            """Compute the limit date for a given date"""
+            if context is None:
+                context = {}
+            if not context.get('product_id', False):
+                date = False
+            else:
+                product = openerp.registry(cr.dbname)['product.product'].browse(
+                    cr, uid, context['product_id'])
+                duration = getattr(product, dtype)
+                # set date to False when no expiry time specified on the product
+                date = duration and (datetime.datetime.today()
+                    + datetime.timedelta(days=duration))
+            return date and date.strftime('%Y-%m-%d %H:%M:%S') or False
+        return calc_date
+
+    _columns = {
+        'life_date': fields.datetime('End of Life Date',
+            help='This is the date on which the goods with this Serial Number may become dangerous and must not be consumed.'),
+        'use_date': fields.datetime('Best before Date',
+            help='This is the date on which the goods with this Serial Number start deteriorating, without being dangerous yet.'),
+        'removal_date': fields.datetime('Removal Date',
+            help='This is the date on which the goods with this Serial Number should be removed from the stock.'),
+        'alert_date': fields.datetime('Alert Date',
+            help="This is the date on which an alert should be notified about the goods with this Serial Number."),
+    }
+    # Assign dates according to products data
+    def create(self, cr, uid, vals, context=None):
+        context = dict(context or {})
+        context['product_id'] = vals.get('product_id', context.get('default_product_id'))
+        return super(stock_production_lot, self).create(cr, uid, vals, context=context)
+
+    _defaults = {
+        'life_date': _get_date('life_time'),
+        'use_date': _get_date('use_time'),
+        'removal_date': _get_date('removal_time'),
+        'alert_date': _get_date('alert_time'),
+    }
