+++ addons/product_expiry/views/product_expiry_view.xml
@@ -20,34 +20,6 @@
             </xpath> 
             </field>
         </record>
-
-        <record model="ir.ui.view" id="view_product_form_expiry">
-            <field name="name">product.normal.form</field>
-            <field name="model">product.template</field>
-            <field name="inherit_id" ref="stock.view_template_property_form" />
-            <field name="arch" type="xml">
-                <group name="stock_property" position="after">
-                    <group name="expiry_and_lots" groups="stock.group_production_lot">
-                        <field name="life_time" />
-                        <field name="use_time" />
-                        <field name="removal_time" />
-                        <field name="alert_time" />
-                    </group>
-                </group>
-            </field>
-        </record>
-
-        <record model="ir.ui.view" id="view_quant_form_expiry">
-            <field name="name">stock.quant.form</field>
-            <field name="model">stock.quant</field>
-            <field name="inherit_id" ref="stock.view_stock_quant_form"/>
-            <field name="arch" type="xml">
-                <xpath expr="//field[@name='lot_id']" position="after" >
-                    <field name="removal_date"/>
-                </xpath>
-            </field>
-        </record>
-
     </data>
 </openerp>
