+++ addons/product_expiry/views/product_template_views.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" ?>
+<openerp>
+    <data>
+        <record model="ir.ui.view" id="view_product_form_expiry">
+            <field name="name">product.normal.form</field>
+            <field name="model">product.template</field>
+            <field name="inherit_id" ref="stock.view_template_property_form" />
+            <field name="arch" type="xml">
+                <group name="stock_property" position="after">
+                    <group name="expiry_and_lots" groups="stock.group_production_lot">
+                        <field name="life_time" />
+                        <field name="use_time" />
+                        <field name="removal_time" />
+                        <field name="alert_time" />
+                    </group>
+                </group>
+            </field>
+        </record>
+</data>
+</openerp>
