+++ addons/product_expiry/views/stock_quant_views.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" ?>
+<openerp>
+    <data>
+        <record model="ir.ui.view" id="view_quant_form_expiry">
+            <field name="name">stock.quant.form</field>
+            <field name="model">stock.quant</field>
+            <field name="inherit_id" ref="stock.view_stock_quant_form"/>
+            <field name="arch" type="xml">
+                <xpath expr="//field[@name='lot_id']" position="after" >
+                    <field name="removal_date"/>
+                </xpath>
+            </field>
+        </record>
+</data>
+</openerp>
